+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 `1.0.20`
 
+- Moved default values from `Sockets` class into `Sockets.Default` name space inner class.
+
 ---
 
 `1.0.19`