+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Changed `OAuth2TokenProvider` to throw `AuthenticationException` if the token client returns a token with invalid `expires_in` field.
 - Added `OAuth2TokenProviderRegistry.getProvider` method to return a registered provider by name.
 - Added variations of the `OAuth2TokenProviderRegistry.of` method without the token provider name function.
+- Added `AlwaysEmptyList` class that represents a list that is always empty no matter what operations are called on it.
 
 ---