+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added variations of the `OAuth2TokenProviderRegistry.of` method without the token provider name function.
 - Added `AlwaysEmptyList` class that represents a list that is always empty no matter what operations are called on it.
 - Added `ErrorObservingInputStream` as a wrapper around an `InputStream` that observes errors.
+- Renamed `BulkCipher._3DES_EDE_CBC` to `BulkCipher.TRIPLE_DES_EDE_CBC` to match Java constant naming conventions.
 
 ---